Add --expected-exit-code argument to the `ios run` command
